backend/requirements.txt
# Core Django
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2

# API Documentation
drf-spectacular==0.27.2

# Database
psycopg2-binary==2.9.9

# HTTP Client
requests==2.31.0
httpx==0.27.0

# LLM Integration
anthropic==0.25.6
openai==1.30.1
google-generativeai==0.8.3  # Google Gemini
# Note: Grok uses OpenAI-compatible API

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Testing
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-asyncio==0.23.6
factory-boy==3.3.0

# Code Quality
black==24.4.2
flake8==7.0.0
isort==5.13.2

# Utilities
python-dotenv==1.0.1
pytz==2024.1
